indexer-service: Fix handling allocations created and mined in different epochs
An allocation that is created with a channel public key _just_ before an
epoch ends will have a channel public key that includes the previous epoch
but will have `createdAtEpoch` set to the new epoch. This was previously
breaking indexer-service. This commit handles this by trying to create
the channel key pair using both epochs. However, if the transaction that
included the allocation takes longer than a full epoch, we can still run
into this problem. We made the error message more useful.