indexer-service: Sign attestations with allocation-specific keys
In order to achieve this, the indexer-service needs to monitor its
active allocations and reconstruct the private keys for its allocations.
It does this using the `allocationSigner` helper introduced in
`indexer-common`.

NOTE: This is a breaking change and requires indexers to re-allocate so
that their agents and services remain aligned, because existing
allocation IDs were created using random inputs.