indexer-agent: Sync data independently, reconcile every 2 minutes
Before, we'd only reconcile if any of the inputs changed, but sometimes
the agent finishes reconciling but should do another pass sooner than
e.g. the next epoch.

This is the case, for instance, if allocations needed to be closed to
free up stake for new allocations; in this case the agent will currently
reconcile allocations per deployment but if allocations of a later
deployment free up stake that could've been used for an earlier
deployment, then we need to reconcile again asap.

On the "asap": It should be possible to flag when there is more
reconciliation work to do, but identifying whether that is even possible
is hard. Simply reconciling again every 2 minutes, without necessarily
resyncing all input data is a decently simple workaround.

This commit therefore polls the different pieces of input data
independently and adds a 2min-timer into the mix to reconcile more
often. This is achieved with a new `join` eventual.