indexer-common: Validate cost models in `setCostModel`
I tried baking this right into the sequelize model so that we
could be certain a model is valid before it hits the db, but
I always got errors that `CostModel` was not set on `undefined`,
so something with the `addon` inside the cost model library wasn't
right. Either way, this does what we want it to do.