*: Wait before collecting receipts for a transfer
This gives indexer-service some time to flush receipts, but the way
this is implemented right now still doesn't guarantee that the
indexer won't receive more queries against the same receipts even
after we've waited for 60s. A better solution is indicated in the
comments. We'll implement something like that next.