*: Ensure allocation summaries for pre-existing allocations
To track allocations and associated receipts, the agent creates allocation
summaries in the db for every allocation it opens. It then links incoming
receipts to these based on the allocation ID. The problem is, these allocation
summaries are only created when allocations are created. Most, if not all
indexers, however, already have allocations from before 0.15.0 and there's
no summaries for these.

This commit ensures allocation summaries are created when flushing receipts, to
make sure receipts that were created before 0.15.0 (or manually without the
agent running) can be flushed to the database.