indexer-agent: Catch subgraph query errors
Some subgraph queries successfully respond with error codes. However,
the return type of the subgraph query was inconsistent with the
structure of the data being returned in the error case. This would
result in error logs that omit the error message because the parsed
response body would result in a type error instead. This change prevents
the type error from occurring such that the log shows the expected error
message.