indexer-service: Add /subgraphs/health/:deployment route
This route handles only GET requests. It returns an HTTP 200 response if
the deployment in question has not failed and is caught up to the chain
head. Otherwise, a 500 is returned with different error messages
depending on what is wrong.

This can be used to check the health of specific subgraph deployments in
environments where the indexing status API is too complicated and
simple, URL-based health checks are the only option (e.g. Cloudflare
load balancers, k8s readiness probes, status pages).