*: Use signalledTokens instead of signalAmount for minSignal threshold
The `signalledTokens` field describes the amount of GRT signalled on a
subgraph or subgraph deployment. `signalAmount` describes the amount of
signal shares on a subgraph, which is based on the bonding curve and
hence don't correspond 1:1 to the amount of GRT signalled.

It is therefore much more intuitive to use `signalledTokens` for the
`minSignal` threshold in indexing rules.