added fantom and polygon to feature support matrix
Proposing adding support for Fantom and Polygon to the feature support matrix. This PR will be included in the GGP (Graph Governance Proposal) to be created here https://snapshot.org/#/council.graphprotocol.eth.

This PR should be merged after Council approval. Afterward, the reference to the canonical GGP-ratified feature support matrix (end of the .md) will be updated.