fix: handle errors fetching actions and vouchers from the DB
This fix adds error handling to the two timers that periodically query
the database: one which gets actions from the queue, and other that gets pending
vouchers. In tests, the DB tables are dropped, so the timers would throw
errors and cause tests to fail. In production whatever errors cause
a DB fetch to fail could be recoverable so the timer will keep retrying
rather than bubbling up an exception, so hopefully that is okay.

We also change CI to only build using Ubuntu 22.04
(the node 18 version has dependency issues on Ubuntu 20.04) and on
node 16 and 18 (the node 17 version was failing to download node, and is EOL).

Additionally we make CI tests and formatting check happen on all PRs, not
only the ones to main.